fix: fixed schema

+++ libs/back/prisma/src/migrations/20250527070507_add_bsda_quantity_refused/migration.sql
@@ -2,5 +2,5 @@
 ALTER TABLE "Bsda" ADD COLUMN     "destinationReceptionRefusedWeight" DECIMAL(65,30);
 
 -- AlterTable
-ALTER TABLE "BsdaRevisionRequest" ADD COLUMN     "destinationReceptionRefusedWeight" DOUBLE PRECISION,
+ALTER TABLE "BsdaRevisionRequest" ADD COLUMN     "destinationReceptionRefusedWeight" DECIMAL(65,30),
 ADD COLUMN     "initialDestinationReceptionRefusedWeight" DECIMAL(65,30);
